improved: refactoring
