+++ src/main/java/exceptions/account/GetAccountByIdException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package exceptions.account;
 
 public class GetAccountByIdException extends RuntimeException {
     public GetAccountByIdException(Throwable cause) {
