+++ src/main/java/exceptions/account/SaveAccountException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package exceptions.account;
 
 public class SaveAccountException extends RuntimeException {
     public SaveAccountException(Throwable cause) {
